+++ CHANGELOG.md
@@ -17,6 +17,62 @@ Here is a template for new release sections
 -
 ```
 ## [Unreleased]
+### Added
+
+### Changed
+
+### Removed
+
+## [0.2.0] - 2020-03-21
+
+Pre-release version of the NESP website
+
+### Added
+- social media links (#74)
+- resource content (#75)
+- meta tags in header (#78)
+- tests of page access (#79)
+- build badge to readme (#82)
+- connection to postgres database (#85)
+- gauge max value is queried from the database (#96)
+- all gauges values are queried from the database (#98)
+- index.py file (#101)
+- objectives route and blueprint (#128)
+- links from logo (#129)
+- about-map page (#153, #162)
+- fadein for features (#155)
+- automatic redraw of gauges when scrolling over (#157)
+- automatic interval change of feature when section is in viewport (#157)
+- link to web development credit page (#157)
+- developed-by page (#162)
+
+### Changed
+- allow use of variable via URL (#76)
+- make table in about responsive and centered (#88)
+- dockerfile setup to use postgres (#91)
+- OG grid mapping image (#95)
+- gauge line is thicker (#96)
+- adapt setup_maps.py for windows users (#101)
+- the webmap uses its own base template (without header and footer) (#101)
+- ratio of giz logos (#129)
+- objective cards design (#145)
+- images size (#148)
+- nigeria se4all section in about page centered with images inline (#154) 
+- rename about-map.html credits.html (#162)
+
+### Removed
+- app/index.py file (#101)
+- outline around the clicked links (#129)
+- about-map blueprint (#162)
+
+### Fixed
+- chrome display bug #64 for small screens (#86)
+- return button was not working (#93)
+- mismatching font for gauges numbers #97 (#98)
+
+## [0.1.0] - 2019-12-05
+
+First version of the NESP website
 
 ### Added
 
@@ -48,6 +104,7 @@ First version of the NESP website
 - page for privacy policy and terms of service (#67)
 - cross-linking protection for all external links (#69)
 - js folder for javascript code (#71)
+- objective page (#132)
 
 ### Changed
 - structure of the folders --> flask files in `app` folder (#12)